+++ proto/client-api/Cargo.toml
@@ -21,7 +21,6 @@ extended-stats = []
 [dependencies]
 derive_more = "0.99"
 medea-macro = { path = "../../crates/medea-macro" }
-medea-reactive = { path = "../../crates/medea-reactive" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_with = "1.4"
