+++ proto/client-api/src/lib.rs
@@ -134,10 +134,7 @@ pub enum Command {
         mids: HashMap<TrackId, String>,
     },
     /// Web Client sends SDP Answer.
-    MakeSdpAnswer {
-        peer_id: PeerId,
-        sdp_answer: String,
-    },
+    MakeSdpAnswer { peer_id: PeerId, sdp_answer: String },
     /// Web Client sends Ice Candidate.
     SetIceCandidate {
         peer_id: PeerId,
@@ -155,9 +152,10 @@ pub enum Command {
         tracks_patches: Vec<TrackPatch>,
     },
 
-    SynchronizeMe {
-        snapshot: RoomSnapshot,
-    },
+    /// Web Client asks media server to synchronize server state and send
+    /// [`Command::SnapshotSynchronized`] with the synchronized
+    /// [`RoomSnapshot`] back to him.
+    SynchronizeMe { snapshot: RoomSnapshot },
 }
 
 /// Web Client's Peer Connection metrics.
@@ -257,10 +255,7 @@ pub enum Event {
 
     /// Media Server notifies Web Client about necessity to apply specified SDP
     /// Answer to Web Client's RTCPeerConnection.
-    SdpAnswerMade {
-        peer_id: PeerId,
-        sdp_answer: String,
-    },
+    SdpAnswerMade { peer_id: PeerId, sdp_answer: String },
 
     /// Media Server notifies Web Client about necessity to apply specified
     /// ICE Candidate.
@@ -271,9 +266,7 @@ pub enum Event {
 
     /// Media Server notifies Web Client about necessity of RTCPeerConnection
     /// close.
-    PeersRemoved {
-        peer_ids: Vec<PeerId>,
-    },
+    PeersRemoved { peer_ids: Vec<PeerId> },
 
     /// Media Server notifies about necessity to update [`Track`]s in specified
     /// Peer.
@@ -285,9 +278,11 @@ pub enum Event {
         tracks_patches: Vec<TrackPatch>,
     },
 
-    SnapshotSynchronized {
-        snapshot: RoomSnapshot,
-    },
+    /// Media Server notifies Web Client that his state should be updated to
+    /// this state.
+    ///
+    /// This [`Event`] is answer on [`Command::SynchronizeMe`].
+    SnapshotSynchronized { snapshot: RoomSnapshot },
 }
 
 /// Represents [RTCIceCandidateInit][1] object.
