+++ proto/client-api/src/snapshots/room.rs
@@ -35,6 +35,12 @@ impl Default for RoomSnapshot {
     }
 }
 
+/// Accessor to the `Room` snapshot objects.
+///
+/// For this trait is implemented `CommandHandler` and
+/// `EventHandler` which will be used on the Web Client side and on the Media
+/// Server side. But real snapshot objects are different on the Web Client and
+/// on the Media Server, so this abstraction is needed.
 pub trait RoomSnapshotAccessor {
     type Peer: PeerSnapshotAccessor;
